requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0